% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{\code{fars_read} function reads a csv file as a dataframe
the function takes a string argurment which is the name of the data filename
if it does not exist, it will generate an error like this: "File 'filename' does not exist"
otherwise, the function will recognise the data and will read the filename as a dataframe}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string character that represents a data filename in csv format to be read}
}
\value{
A data frame read from the filename argument using the dplyr package. If it does not exists, it will throw an error
}
\description{
\code{fars_read} function reads a csv file as a dataframe
the function takes a string argurment which is the name of the data filename
if it does not exist, it will generate an error like this: "File 'filename' does not exist"
otherwise, the function will recognise the data and will read the filename as a dataframe
}
\examples{
\dontrun{fars_read("accident_2015.csv.bz2")}

}
