% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{\code{fars_read_years} function allows to read a vector of years
the function uses the lapply function and the two functions created above
to create a new csv filename for each year and then test if the filename exists
then by using the dplyr package only 2 columns will be visible: month and year
if the year is invalid, it will return "NULL"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years to be read}
}
\value{
returns a list of dataframes with only the month and year columns
}
\description{
\code{fars_read_years} function allows to read a vector of years
the function uses the lapply function and the two functions created above
to create a new csv filename for each year and then test if the filename exists
then by using the dplyr package only 2 columns will be visible: month and year
if the year is invalid, it will return "NULL"
}
\examples{
\dontrun{fars_read_years(c(2013, 2014, 2015))}

}
