% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{\code{fars_summarize_years} function allows to summarise the data by years
the function takes a list of vectors (years) which are passed in the third function above
then dataframes are binded to be summarised the number of cases per year
from a longitudinal format to wide format is converted by using the spead function}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years to be read}
}
\value{
returns a dataframe with the number of cases summarised by year in a wide format
}
\description{
\code{fars_summarize_years} function allows to summarise the data by years
the function takes a list of vectors (years) which are passed in the third function above
then dataframes are binded to be summarised the number of cases per year
from a longitudinal format to wide format is converted by using the spead function
}
\examples{
\dontrun{fars_summarize_years(c(2013, 2014, 2015))}

}
