% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{\code{make_filename} function allows to create a filename
the function takes a numeric or string argument (a year) and a csv filename is created
based on that year. the function guarantees that the argument be converted to as integer}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric or string argument which is the year for the future filename}
}
\value{
A combination of text and the year argument which yields the csv filename as a text type
}
\description{
\code{make_filename} function allows to create a filename
the function takes a numeric or string argument (a year) and a csv filename is created
based on that year. the function guarantees that the argument be converted to as integer
}
\examples{
\dontrun{make_filename(2015)}

}
